requirements.txt
pytorch-lightning==2.0.0
pytorch==2.2.2
torchinfo==1.8.0
transformers==4.40.2
pandas==2.2.2
scikit-learn==1.4.2
wandb==0.17.4
av==12.0.0
h5py==3.11.0
sox==1.5.0
soundfile==0.12.1
torchaudio==2.2.2
torchvision==0.17.2
onnx==1.16.0
onnxscript==0.1.0.dev20241022
librosa==0.9.2
numba==0.56.2
GPUtil==1.4.0